refactor(evaluations): PreviewEvaluation sous 300 lignes — éclatement complet (H1)
PreviewEvaluation.vue 789 → 244 lignes. Éclatement façon G1 avec CSS distribué
VERBATIM dans chaque sous-composant :
- usePreviewEvaluation.js (105) : chargement preview, réponses simulées +
  progression, publication/édition/retour.
- PreviewStatsGrid.vue (111) : grille de stats (durée/questions/points/barème).
- PreviewProgress.vue (58) : barre de progression.
- PreviewQuestionsList.vue (236) : questions QCM + réponse courte (+ CSS options,
  inputs, animation fadeIn).
- PreviewActionsFooter.vue (145) : pied d'actions (retour/édition/publication).

L'orchestrateur conserve header+bandeau+état d'erreur et leur CSS résiduel.
Services, routes, logs, confirm et alertes identiques. Tous fichiers ≤300.

Tests : 12 verts (composable + 4 montages). Build vert.
